app/User panel/cloud_functions/gcp_fetch_student_credentials_data/requirements.txt
google-cloud-storage==2.8.0

